test(dotnet-system): rename adapter ServicesTest back to LifeCycleTest
The 2021 "Lifecycle -> Services" sweep (12020547) blanket-renamed
LifeCycleTest.cs to ServicesTest.cs at 99% similarity, although the class
tests object/transaction lifecycle (NextId, Create, Delete, IsDeleted,
Versioning, Rollback, Instantiate, prefetch) and contains no
service-container tests. Restore the original, accurate name across the 4
Tests.Static files (base + Memory/Npgsql/SqlClient), matching the Workspace
suite which already uses LifecycleTests. The genuine *Services production
interfaces (IDatabaseServices, ISessionServices, ...) are correctly named
and left untouched.
