+++ dotnet/System/Database/Adapters/Tests.Static/Static/Memory/LifeCycleTest.cs
@@ -1,24 +1,30 @@
-// <copyright file="ServicesTest.cs" company="Allors bv">
+// <copyright file="LifeCycleTest.cs" company="Allors bv">
 // Copyright (c) Allors bv. All rights reserved.
 // Licensed under the LGPL license. See LICENSE file in the project root for full license information.
 // </copyright>
 
-namespace Allors.Database.Adapters.Sql.Npgsql
+namespace Allors.Database.Adapters.Memory
 {
+    using System;
     using Adapters;
     using Xunit;
 
-    public class ServicesTest : Adapters.ServicesTest, IClassFixture<Fixture<ServicesTest>>
+    public class LifeCycleTest : Adapters.LifeCycleTest, IDisposable
     {
-        private readonly Profile profile;
-
-        public ServicesTest() => this.profile = new Profile(this.GetType().Name);
+        private readonly Profile profile = new Profile();
 
         protected override IProfile Profile => this.profile;
 
+        [Fact]
+        public override void DifferentTransactions()
+        {
+        }
+
         public override void Dispose() => this.profile.Dispose();
 
-        protected override void SwitchDatabase() => this.profile.SwitchDatabase();
+        protected override void SwitchDatabase()
+        {
+        }
 
         protected override IDatabase CreatePopulation() => this.profile.CreateDatabase();
