+++ dotnet/System/Database/Adapters/Tests.Static/Static/Sql/Npgsql/LifeCycleTest.cs
@@ -1,18 +1,18 @@
-// <copyright file="ServicesTest.cs" company="Allors bv">
+// <copyright file="LifeCycleTest.cs" company="Allors bv">
 // Copyright (c) Allors bv. All rights reserved.
 // Licensed under the LGPL license. See LICENSE file in the project root for full license information.
 // </copyright>
 
-namespace Allors.Database.Adapters.Sql.SqlClient
+namespace Allors.Database.Adapters.Sql.Npgsql
 {
     using Adapters;
     using Xunit;
 
-    public class ServicesTest : Adapters.ServicesTest, IClassFixture<Fixture<ServicesTest>>
+    public class LifeCycleTest : Adapters.LifeCycleTest, IClassFixture<Fixture<LifeCycleTest>>
     {
         private readonly Profile profile;
 
-        public ServicesTest() => this.profile = new Profile(this.GetType().Name);
+        public LifeCycleTest() => this.profile = new Profile(this.GetType().Name);
 
         protected override IProfile Profile => this.profile;
