+++ dotnet/System/Database/Adapters/Tests.Static/Static/Sql/SqlClient/LifeCycleTest.cs
@@ -1,30 +1,24 @@
-// <copyright file="ServicesTest.cs" company="Allors bv">
+// <copyright file="LifeCycleTest.cs" company="Allors bv">
 // Copyright (c) Allors bv. All rights reserved.
 // Licensed under the LGPL license. See LICENSE file in the project root for full license information.
 // </copyright>
 
-namespace Allors.Database.Adapters.Memory
+namespace Allors.Database.Adapters.Sql.SqlClient
 {
-    using System;
     using Adapters;
     using Xunit;
 
-    public class ServicesTest : Adapters.ServicesTest, IDisposable
+    public class LifeCycleTest : Adapters.LifeCycleTest, IClassFixture<Fixture<LifeCycleTest>>
     {
-        private readonly Profile profile = new Profile();
+        private readonly Profile profile;
+
+        public LifeCycleTest() => this.profile = new Profile(this.GetType().Name);
 
         protected override IProfile Profile => this.profile;
 
-        [Fact]
-        public override void DifferentTransactions()
-        {
-        }
-
         public override void Dispose() => this.profile.Dispose();
 
-        protected override void SwitchDatabase()
-        {
-        }
+        protected override void SwitchDatabase() => this.profile.SwitchDatabase();
 
         protected override IDatabase CreatePopulation() => this.profile.CreateDatabase();
